Move logic to count visible elements to selenium helper
